+++ src/slidescene.cpp
@@ -1,10 +1,12 @@
-// SPDX-FileCopyrightText: 2022 Valentin Bruch <software@vbruch.eu>
+// SPDX-FileCopyrightText: 2023 Valentin Bruch <software@vbruch.eu>
 // SPDX-License-Identifier: GPL-3.0-or-later OR AGPL-3.0-or-later
 
 #include <iterator>
 #include <algorithm>
 #include <QTransform>
 #include <QString>
+#include <QSvgGenerator>
+#include <QSvgRenderer>
 #include <QByteArray>
 #include <QDataStream>
 #include <QGraphicsVideoItem>
@@ -37,6 +39,7 @@
 #include "src/drawing/ellipsegraphicsitem.h"
 #include "src/drawing/linegraphicsitem.h"
 #include "src/drawing/arrowgraphicsitem.h"
+#include "src/drawing/graphicspictureitem.h"
 #include "src/drawing/pointingtool.h"
 #include "src/drawing/texttool.h"
 #include "src/drawing/selectiontool.h"
@@ -1823,136 +1826,30 @@ void SlideScene::updateSearchResults()
 
 void readFromSVG(const QByteArray &data, QList<QGraphicsItem*> &target)
 {
-    QXmlStreamReader reader(data);
-    while (!reader.atEnd() && (reader.readNext() != QXmlStreamReader::StartElement || reader.name().toUtf8() != "svg")) {}
-    while (reader.readNextStartElement())
-    {
-        debug_msg(DebugDrawing, reader.name());
-        if (reader.name().toUtf8() == "g")
-        {
-            const QPointF pos(reader.attributes().value("x", "0.").toDouble(), reader.attributes().value("y", "0.").toDouble());
-            QTransform transform;
-            QString transform_string = reader.attributes().value("transform").toString();
-            if (transform_string.length() >= 19 && transform_string.startsWith("matrix(") && transform_string.endsWith(")"))
-            {
-                // TODO: this is inefficient
-                transform_string.remove("matrix(");
-                transform_string.remove(")");
-                const QStringList list = transform_string.trimmed().replace(" ", ",").split(",");
-                if (list.length() == 6)
-                    transform.setMatrix(list[0].toDouble(), list[1].toDouble(), 0., list[2].toDouble(), list[3].toDouble(), 0., list[4].toDouble(), list[5].toDouble(), 1.);
-            }
-            while (reader.readNextStartElement() && reader.name().toUtf8() != "path" && !reader.isEndElement())
-                reader.skipCurrentElement();
-            debug_msg(DebugDrawing, reader.name());
-            if (reader.name().toUtf8() != "path")
-                continue;
-
-            QString coordinates = reader.attributes().value("d").toString();
-            coordinates.replace(",", " ");
-            // TODO: currently all coordinates are interpreted as absolute coordinates. This is wrong!
-            coordinates.remove("m");
-            coordinates.remove("l");
-            coordinates.remove("z");
-            coordinates.remove("L");
-            coordinates.remove("M");
-            QRegularExpression re;
-            re.setPattern(R"(\s\s+)");
-            coordinates.replace(re, " ");
-            coordinates = coordinates.trimmed();
-            QString style = reader.attributes().value("style").toString();
-            QColor color = Qt::black;
-            qreal width = 1.;
-            re.setPattern("stroke:(#[0-9a-fA-F]+)");
-            if (style.contains("stroke:"))
-                color = QColor(re.match(style).captured(1));
-            re.setPattern("stroke-width:([0-9.]+)");
-            if (style.contains("stroke-width:"))
-                width = re.match(style).captured(1).toDouble();
-            DrawTool tool(Tool::BasicSelectionTool, 0, QPen(color, width));
-            BasicGraphicsPath *path = new BasicGraphicsPath(tool, coordinates);
-            path->setPos(pos);
-            path->setTransform(transform);
-            target.append(path);
-            if (!reader.isEndElement())
-                reader.skipCurrentElement();
-        }
-        else if (reader.name().toUtf8() == "text")
-        {
-            // TODO
-        }
-        if (!reader.isEndElement())
-            reader.skipCurrentElement();
-    }
-    reader.clear();
+    QSvgRenderer renderer(data);
+    QPicture picture;
+    QPainter painter;
+    if (!painter.begin(&picture))
+        return;
+    renderer.render(&painter);
+    painter.end();
+    target.append(new GraphicsPictureItem(picture));
 }
 
 void writeToSVG(QByteArray &data, const QList<QGraphicsItem*> &source, const QRectF &rect)
 {
-    QXmlStreamWriter writer(&data);
-    writer.setAutoFormatting(true);
-    writer.setAutoFormattingIndent(0);
-    writer.writeStartDocument();
-    writer.writeComment(" Created with BeamerPresenter ");
-    writer.writeStartElement("svg");
-    writer.writeAttribute("version", "1.1");
-    writer.writeAttribute("xmlns", "http://www.w3.org/2000/svg");
-    //writer.writeAttribute("width", QString::number(rect.width()));
-    //writer.writeAttribute("height", QString::number(rect.height()));
-    writer.writeAttribute("viewBox", QString("%1 %2 %3 %4").arg(rect.x()).arg(rect.y()).arg(rect.width()).arg(rect.height()));
-    for (auto &item : source)
+    QSvgGenerator generator;
+    QBuffer buffer(&data);
+    generator.setOutputDevice(&buffer);
+    generator.setViewBox(rect);
+    QPainter painter;
+    if (!painter.begin(&generator))
+        return;
+    QStyleOptionGraphicsItem style;
+    for (const auto item : source)
     {
-        const QTransform transform = item->transform();
-        QString matrix = QString("matrix(%1,%2,%3,%4,%5,%6)")
-                .arg(transform.m11())
-                .arg(transform.m12())
-                .arg(transform.m21())
-                .arg(transform.m22())
-                .arg(transform.dx())
-                .arg(transform.dy());
-        switch (item->type())
-        {
-        case BasicGraphicsPath::Type:
-        case FullGraphicsPath::Type:
-        {
-            // TODO: flexible width of FullGraphicsPath is ignored.
-            const AbstractGraphicsPath *path = static_cast<const AbstractGraphicsPath*>(item);
-            writer.writeStartElement("g");
-            writer.writeAttribute("x", QString::number(item->x()));
-            writer.writeAttribute("y", QString::number(item->y()));
-            writer.writeAttribute("transform", matrix);
-            writer.writeStartElement("path");
-            const DrawTool &tool = path->getTool();
-            QString style = "fill:";
-            style += tool.brush().style() == Qt::NoBrush ? "none" : tool.brush().color().name();
-            style += ";stroke:";
-            style += tool.color().name();
-            style += ";stroke-width:";
-            style += QString::number(tool.width());
-            style += ";stroke-linecap:round;stroke-linejoin:round;";
-            writer.writeAttribute("style", style);
-            writer.writeAttribute("d", path->svgCoordinates());
-            writer.writeEndElement();// "path" element
-            writer.writeEndElement(); // "g" element
-            break;
-        }
-        case TextGraphicsItem::Type:
-        {
-            const TextGraphicsItem *textItem = static_cast<const TextGraphicsItem*>(item);
-            writer.writeStartElement("text");
-            const QPointF origin = item->sceneBoundingRect().bottomLeft();
-            // TODO: coordinates are shifted.
-            writer.writeAttribute("x", QString::number(origin.x()));
-            writer.writeAttribute("y", QString::number(origin.y()));
-            writer.writeAttribute("transform", matrix);
-            writer.writeAttribute("style", "font-size:" + QString::number(textItem->font().pointSizeF()) + "pt");
-            // TODO: currently only plain text is supported.
-            writer.writeCharacters(textItem->toPlainText());
-            writer.writeEndElement();
-            break;
-        }
-        }
+        painter.setTransform(item->sceneTransform(), false);
+        item->paint(&painter, &style);
     }
-    writer.writeEndElement(); // "svg" element
-    writer.writeEndDocument();
+    painter.end();
 }